requirements.txt
# Web framework
fastapi
uvicorn[standard]
pydantic>=2
python-multipart
aiofiles
tenacity

# OpenAI for embeddings and STT
openai>=1.51.0

# Supabase authentication
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0

# Vector search
faiss-cpu

# Database
psycopg2-binary
SQLAlchemy>=2.0
python-dotenv

# AWS S3 (for future use)
boto3

# NumPy for FAISS
numpy
gunicorn
